perf: lazy load password reset components
